+++ reapy/core/__init__.py
@@ -4,7 +4,9 @@ from .audio_accessor import AudioAccessor
 from .envelope import Envelope, EnvelopeList, EnvelopePoint
 from .fx import FX, FXList, FXParam, FXParamsList
 from .item import CC, CCList, Item, Note, NoteList, Source, Take
-from .project import Marker, Project, Region, TimeSelection
+from .project import (
+    Marker, MarkerInfo, Project, Region, RegionInfo, TimeSelection
+)
 from .track import AutomationItem, Send, Track, TrackList
 from .window import MIDIEditor, ToolTip, Window
 
@@ -34,8 +36,10 @@ __all__ = [
     "Take",
     # core.project
     "Marker",
+    "MarkerInfo",
     "Project",
     "Region",
+    "RegionInfo",
     "TimeSelection",
     # core.track
     "AutomationItem",