+++ reapy/core/project/__init__.py
@@ -1,4 +1,4 @@
-from .marker import Marker
+from .marker import Marker, MarkerInfo
 from .project import Project
-from .region import Region
+from .region import Region, RegionInfo
 from .time_selection import TimeSelection